requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
appdirs==1.4.4
arrow==1.3.0
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.3.0
billiard==4.2.2
binaryornot==0.4.4
blis==1.3.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
contractions==0.1.73
cookiecutter==2.6.0
cryptography==46.0.1
cymem==2.0.11
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.2
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
frozenlist==1.7.0
fsspec==2025.9.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.8.0
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.4
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.3
omegaconf==2.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.4.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
pyahocorasick==2.2.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.2
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqltrie==0.11.2
srsly==2.5.1
tabulate==0.9.0
text-unidecode==1.3
textsearch==0.0.24
thinc==8.3.6
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==4.0
